# docs/swagger.yaml
definitions:
  syncer.Response:
    properties:
      code:
        type: integer
      data: {}
      detail:
        type: string
      msg:
        type: string
    type: object
host: 127.0.0.1:8080
info:
  contact:
    name: API Support
    url: http://www.swagger.io/support
  description: 区块链浏览器后端API接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: 区块链浏览器
  version: "1.0"
paths:
...
  /peers:
    get:
      consumes:
      - application/json
      description: 分页获取节点列表
      parameters:
      - description: 页码
        in: query
        name: page_num
        type: integer
      - description: 个数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code":!200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 节点列表
      tags:
      - 节点
  /peers/{name}:
    get:
      consumes:
      - application/json
      description: 获取节点信息
      parameters:
      - description: 节点ID
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 节点信息
      tags:
      - 节点
...
  /transactions:
    get:
      consumes:
      - application/json
      description: 分页获取交易列表
      parameters:
      - description: 页码
        in: query
        name: page_num
        type: integer
      - description: 个数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code":!200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 交易列表
      tags:
      - 交易
  /transactions/{name}:
    get:
      consumes:
      - application/json
      description: 获取交易信息
      parameters:
      - description: 交易哈希
        in: path
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 交易信息
      tags:
      - 交易
  /validators:
    get:
      consumes:
      - application/json
      description: 分页获验证者列表
      parameters:
      - description: 页码
        in: query
        name: page_num
        type: integer
      - description: 个数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code":!200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 验证者列表
      tags:
      - 验证者
  /validators/{name}:
    get:
      consumes:
      - application/json
      description: 获取验证者信息
      parameters:
      - description: 账户地址
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 验证者信息
      tags:
      - 验证者
swagger: "2.0"